refactor(computer): move input driver to separate module
